backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
alembic>=1.10.0
redis>=4.6.0
greenlet>=2.0.0
python-socketio>=5.9.0
python-engineio>=4.7.0
aioredis>=2.0.0